test/requirements.txt
pytest==8.1.1
cocotb==1.8.1
numpy==1.24.4
matplotlib==3.7.5
pillow==10.3.0